requirements.txt
apache-airflow-providers-google
pyarrow
paho.mqtt
shed==0.9.7
python-dotenv
